+++ h2datatype/src/simple/network/ipv6.rs
@@ -28,7 +28,7 @@ impl IPv6 {
 }
 
 impl H2TypeTrait for IPv6 {
-    fn base_size(&self, _context: Context) -> SimpleResult<u64> {
+    fn base_size(&self, _context: Context) -> SimpleResult<usize> {
         Ok(16)
     }
