+++ h2datatype/src/simple/network/mac_address.rs
@@ -25,7 +25,7 @@ impl MacAddress {
 }
 
 impl H2TypeTrait for MacAddress {
-    fn base_size(&self, _context: Context) -> SimpleResult<u64> {
+    fn base_size(&self, _context: Context) -> SimpleResult<usize> {
         Ok(6)
     }
