+++ h2datatype/src/simple/numeric/h2float.rs
@@ -38,8 +38,8 @@ impl H2Float {
 }
 
 impl H2TypeTrait for H2Float {
-    fn base_size(&self, _context: Context) -> SimpleResult<u64> {
-        Ok(self.reader.size() as u64)
+    fn base_size(&self, _context: Context) -> SimpleResult<usize> {
+        Ok(self.reader.size())
     }
 
     fn to_display(&self, context: Context) -> SimpleResult<String> {
